level1 - renamed class HolderKeys.kt to KeysHolder. A little correction according to Ktlint
